+++ src/app/Scenes/SubmitArtwork/ArtworkDetails/ArtworkDetails.tests.tsx
@@ -9,7 +9,7 @@ import { RelayEnvironmentProvider } from "react-relay"
 import { act } from "react-test-renderer"
 import { useTracking } from "react-tracking"
 import { createMockEnvironment } from "relay-test-utils"
-import { createConsignSubmission, updateConsignSubmission } from "../Mutations"
+import { createConsignSubmission, updateConsignSubmission } from "../utils/mutations"
 import { ArtworkDetails } from "./ArtworkDetails"
 import { createOrUpdateSubmission } from "./utils/createOrUpdateSubmission"
 import { mockFormValues } from "./utils/testUtils"
