+++ src/app/Scenes/SubmitArtwork/ArtworkDetails/ArtworkDetails.tsx
@@ -6,7 +6,7 @@ import { Formik } from "formik"
 import { BulletedItem, CTAButton, Flex, LinkText, Spacer } from "palette"
 import React, { useState } from "react"
 import { useTracking } from "react-tracking"
-import { ErrorView } from "../Components/ErrorView"
+import { ErrorView } from "../ErrorView/ErrorView"
 import { ArtworkDetailsForm } from "./ArtworkDetailsForm"
 import { createOrUpdateSubmission } from "./utils/createOrUpdateSubmission"
 import { ArtworkDetailsFormModel, artworkDetailsValidationSchema } from "./validation"
