+++ src/app/Scenes/SubmitArtwork/ArtworkDetails/InfoModal/InfoModal.tests.tsx
