+++ src/app/Scenes/SubmitArtwork/ArtworkDetails/InfoModal/InfoModal.tsx
