+++ src/app/Scenes/SubmitArtwork/ArtworkDetails/utils/createOrUpdateSubmission.ts
@@ -3,7 +3,7 @@ import {
   CreateSubmissionMutationInput,
 } from "__generated__/createConsignmentSubmissionMutation.graphql"
 import { UpdateSubmissionMutationInput } from "__generated__/updateConsignSubmissionMutation.graphql"
-import { createConsignSubmission, updateConsignSubmission } from "../../Mutations"
+import { createConsignSubmission, updateConsignSubmission } from "../../utils/mutations"
 import { ArtworkDetailsFormModel } from "../validation"
 import { limitedEditionValue } from "./rarityOptions"
