+++ src/app/Scenes/SubmitArtwork/utils/mutations/removeAssetFromConsignmentSubmissionMutation.ts
