+++ src/app/Scenes/SubmitArtwork/utils/submissionModelState.tsx
